+++ packages/transfer_manager_platform_interface/lib/src/models.dart
@@ -1,6 +1,7 @@
 enum PlatformTaskState {
   enqueued,
   running,
+  paused,
   succeeded,
   failed,
   cancelled,